AGNKCS-7: rename helper
